AI-Models/Creation-Models/style-transfer/requirements.txt
fastapi
uvicorn
python-multipart
pillow
tensorflow>=2.12.0
tensorflow_hub>=0.12.0
numpy